# docs/openapi.yaml
openapi: 3.0.3
info:
  title: The Ultimate API
  description: |
    # 💀 The Ultimate API 
    
    Welcome to **The Ultimate API** - the most comprehensive and powerful Roblox audio downloading service available. This API provides seamless access to Roblox's vast audio library with advanced features including batch processing, detailed analytics, user management, and robust rate limiting.
    
    ## 🚀 Key Features
    
    - **🎵 Audio Downloads**: Download any public Roblox audio file instantly
    - **📦 Batch Processing**: Download multiple audio files in a single request
    - **👥 User Management**: Secure authentication with JWT tokens and API keys
    - **📊 Advanced Analytics**: Detailed statistics and usage tracking
    - **🛡️ Rate Limiting**: Smart rate limiting to prevent abuse
    - **⚡ High Performance**: Async processing for maximum speed
    - **🔒 Secure**: Industry-standard security practices
    
    ## 🛠️ Getting Started
    
    1. **Register** a new account using the `/auth/register` endpoint
    2. **Login** to get your JWT token via `/auth/login`
    3. **Generate** an API key (optional) via `/auth/api-key`
    4. **Start downloading** audio files using `/audio/download`
    
    ## 🔐 Authentication
    
    This API supports two authentication methods:
    - **JWT Tokens**: Perfect for web applications and temporary access
    - **API Keys**: Ideal for server-to-server communication and long-term access
    
    ## 📈 Rate Limits
    
...
    
...
    
    ```
...
servers:
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register New User
      description: |
        Create a new user account on The Ultimate API.
        
        **Registration includes:**
        - Secure password hashing
        - Email validation
        - Username uniqueness check
        - Automatic profile creation
        
        **Requirements:**
        - Username: 3-50 characters, alphanumeric, dashes, underscores only
        - Email: Valid email address format
        - Password: Minimum 8 characters with complexity requirements
        
        **After registration:**
        - Account is immediately active
        - Use `/auth/login` to get your JWT token
        - Generate API keys via `/auth/api-key`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
            examples:
              user_registration:
                summary: Standard User Registration
                value:
                  username: "audio_lover_2025"
                  email: "user@example.com"
                  password: "SecureP@ssw0rd!"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                created_user:
                  summary: Successfully Created User
                  value:
                    id: 1
                    username: "audio_lover_2025"
                    email: "user@example.com"
                    is_active: true
                    is_admin: false
                    total_commands: 0
                    total_downloads: 0
                    successful_downloads: 0
                    failed_downloads: 0
                    created_at: "2025-07-13T00:00:00Z"
                    last_seen: null
        '400':
          description: Validation error or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                username_taken:
                  summary: Username Already Exists
                  value:
                    detail: "Username already registered"
                    error_code: "USER_EXISTS"
                invalid_email:
                  summary: Invalid Email Format
                  value:
                    detail: "Invalid email format"
                    error_code: "VALIDATION_ERROR"
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User Login
      description: |
        Authenticate a user and receive a JWT access token.
        
        **Login Process:**
        - Validates username and password
        - Returns JWT token for authentication
        - Updates last seen timestamp
        - Provides token expiration information
        
        **Token Usage:**
        - Include in Authorization header: `Bearer <token>`
        - Valid for 24 hours by default
        - Use for all authenticated endpoints
        
        **Security Features:**
        - Password hashing verification
        - Rate limiting protection
        - Login attempt tracking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
            examples:
              user_login:
                summary: Standard Login
                value:
                  username: "audio_lover_2025"
                  password: "SecureP@ssw0rd!"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
              examples:
                login_success:
                  summary: Successful Login
                  value:
                    access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    token_type: "bearer"
                    expires_in: 86400
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalid_credentials:
                  summary: Wrong Username/Password
                  value:
                    detail: "Invalid username or password"
                    error_code: "INVALID_CREDENTIALS"
...
  /stats/user:
    get:
      tags:
        - Statistics
      summary: User Statistics
      description: |
        Get comprehensive statistics for the authenticated user.
        
        **Statistics Include:**
        - Total downloads and success rate
        - Bandwidth usage
        - Popular download times
        - File type preferences
        - Monthly/weekly trends
        
        **Metrics Provided:**
        - Download count and success rate
        - Total bandwidth consumed
        - Average file size
        - Peak usage periods
        - Favorite audio categories
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: User statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStats'
  /stats/global:
    get:
      tags:
        - Statistics
      summary: Global Statistics
      description: |
        Get global platform statistics (admin only).
        
        **Global Metrics:**
        - Total registered users
        - Platform-wide downloads
        - Popular audio files
        - System performance metrics
        - Usage trends and patterns
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Global statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalStats'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token authentication. Include the token in the Authorization header:
        ```
        Authorization: Bearer <your_jwt_token>
        ```
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key authentication. Include your API key in the X-API-Key header:
        ```
        X-API-Key: uk_live_your_api_key_here
        ```
  schemas:
    UserCreate:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          pattern: "^[a-zA-Z0-9_-]+$"
          minLength: 3
          maxLength: 50
          description: Unique username (3-50 characters, alphanumeric, dashes, underscores only)
          example: "audio_lover_2025"
        email:
          type: string
          format: email
          description: Valid email address
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          maxLength: 100
          description: Strong password (minimum 8 characters)
          example: "SecureP@ssw0rd!"
    UserLogin:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: Your username
          example: "audio_lover_2025"
        password:
          type: string
          description: Your password
          example: "SecureP@ssw0rd!"
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          description: Unique user ID
          example: 1
        username:
          type: string
          description: Username
          example: "audio_lover_2025"
        email:
          type: string
          format: email
          description: User email address
          example: "user@example.com"
        is_active:
          type: boolean
          description: Account status
          example: true
        is_admin:
          type: boolean
          description: Admin privileges
          example: false
        total_commands:
          type: integer
          description: Total API commands executed
          example: 156
        total_downloads:
          type: integer
          description: Total download attempts
          example: 142
        successful_downloads:
          type: integer
          description: Successful downloads
          example: 138
        failed_downloads:
          type: integer
          description: Failed downloads
          example: 4
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2025-07-13T00:00:00Z"
        last_seen:
          type: string
          format: date-time
          nullable: true
          description: Last activity timestamp
          example: "2025-07-13T12:30:00Z"
    Token:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          description: Token type
          example: "bearer"
        expires_in:
          type: integer
          description: Token expiration time in seconds
          example: 86400
...
    UserStats:
      type: object
      properties:
        total_commands:
          type: integer
          description: Total API commands executed
          example: 156
        total_downloads:
          type: integer
          description: Total download attempts
          example: 142
        successful_downloads:
          type: integer
          description: Successful downloads
          example: 138
        failed_downloads:
          type: integer
          description: Failed downloads
          example: 4
        success_rate:
          type: number
          description: Success rate percentage
          example: 97.18
        first_seen:
          type: string
          format: date-time
          description: First API usage
          example: "2025-07-01T00:00:00Z"
        last_seen:
          type: string
          format: date-time
          nullable: true
          description: Last API usage
          example: "2025-07-13T12:30:00Z"
    GlobalStats:
      type: object
      properties:
        total_users:
          type: integer
          example: 1250
        total_downloads:
          type: integer
          example: 45678
        total_bandwidth:
          type: integer
          description: Total bandwidth in bytes
          example: 1234567890
        popular_files:
          type: array
          items:
            type: object
            properties:
              asset_id:
                type: integer
                example: 142376088
              downloads:
                type: integer
                example: 1234
              title:
                type: string
                example: "Oof Sound Effect"
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error description
          example: "Asset ID not found"
        error_code:
          type: string
          description: Machine-readable error code
          example: "ASSET_NOT_FOUND"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2025-07-13T12:30:00Z"